requirements.txt
asgiref==3.7.2
attrs==23.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Django==5.0.6
django-filter==24.2
django-humanize==0.1.2
django-jalali-date==1.1.2
django-render-partial==0.4
humanize==4.9.0
idna==3.7
importlib-metadata==6.7.0
isodate==0.6.1
jdatetime==4.1.1
lxml==5.2.2
mysql-connector-python==8.0.33
mysqlclient==2.2.4
packaging==24.0
pillow==10.3.0
platformdirs==4.2.2
protobuf==3.20.3
pytimeparse2==1.7.1
pytz==2024.1
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
setuptools==68.0.0
simplejson==3.19.2
six==1.16.0
sorl-thumbnail==12.9.0
sqlite3-to-mysql==2.0.3
sqlparse==0.4.4
sweetify==2.3.1
tabulate==0.9.0
tqdm==4.66.4
typing_extensions==4.7.1
tzdata==2024.1
Unidecode==1.3.8
urllib3==2.2.2
wheel==0.41.2
zeep==4.2.1
zipp==3.15.0
